Move receipt printer value types to Hardware
Dedup ReceiptPrinterDevice and the connection status into the Hardware
module and re-export them through Yosemite typealiases, so POS consumes
Yosemite types (mirrors CardReader). Drop the duplicated `connecting`
from PrinterDiscoveryState and add an `idle` start state; add `idle` to
ReceiptPrinterConnectionStatus as the initial state.
