+++ Modules/Sources/Yosemite/Model/Model.swift
@@ -263,6 +263,8 @@ public typealias CardPresentCaptureMethod = Hardware.CardPresentCaptureMethod
 public typealias RefundParameters = Hardware.RefundParameters
 public typealias PaymentIntent = Hardware.PaymentIntent
 public typealias PrintingResult = Hardware.PrintingResult
+public typealias ReceiptPrinterDevice = Hardware.ReceiptPrinterDevice
+public typealias ReceiptPrinterConnectionStatus = Hardware.ReceiptPrinterConnectionStatus
 public typealias CardPresentReceiptParameters = Hardware.CardPresentReceiptParameters
 public typealias CardPresentTransactionDetails = Hardware.CardPresentTransactionDetails
 public typealias StripeAccount = Networking.StripeAccount
